+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Note: Spacedock's markdown doesn't recognize lists using `-`, so make sure to
 ## Unreleased
 ## Fixed
 * Fixed a bug where empty SRBs would occasionally keep running in the background.
+* Fixed logspam due to missing simulation resources.
 
 ## 0.6.6
 ### Fixed